[OPENBUS-1719] Experimentar o uso do PUTS com outros projetos como DataService e FTC - Ajustes no build/cmake para deixar de usar o TMPDIR como área de build e   criar uma pasta build dentro da pasta com fontes. Uma variação dessa   implementação é chamar a copy.run chamando src_dir e no descritor manter o uso   da TMPDIR como alternativa. A questão é: chamar a copy.run passando a variavel   src_dir (tendo ela o valor da TMPDIR) é que é ruim. - Melhoria no suporte a link simbolicos no util.lua e build/copy.lua
git-svn-id: https://subversion.tecgraf.puc-rio.br/engdist/openbus/puts/trunk@131940 ae0415b3-e90b-0410-900d-d0be9363c56b

+++ lua/tools/util.lua
@@ -49,7 +49,9 @@ function fs.is_dir(at)
    return (at and (os.execute(myplat.cmd.test.." -d "..at) == 0))
 end
 function fs.is_file(at)
-   return (at and (os.execute(myplat.cmd.test.." -f "..at) == 0))
+   return (at and ((os.execute(myplat.cmd.test.." -f "..at) == 0) or 
+                   (os.execute(myplat.cmd.test.." -e "..at) == 0) or
+                   (os.execute(myplat.cmd.test.." -L "..at) == 0)))
 end
 
 function fs.list_dir(at)
@@ -141,12 +143,14 @@ function link(package, orig, linkpath)
     if not cache[package] then cache[package] = { } end
     cache[package].links = assert(io.open(config.PKGDIR.."/"..package..".links", "w"))
   end
-  cache[package].links:write(linkpath.."\n")
   local dir,name = linkpath:gmatch("(.*%/+)(.+)")()
   dir = dir or "."
-  os.execute("cd "..config.INSTALL.TOP .."; ".. myplat.cmd.mkdir .. dir)
-  -- ... and real link to destination
-  os.execute("ln -sf "..orig.." "..config.INSTALL.TOP.."/"..linkpath)
+  if fs.is_file(path.pathname(config.INSTALL.TOP,orig)) then
+    os.execute("cd "..config.INSTALL.TOP .."; ".. myplat.cmd.mkdir .. dir)
+    -- ... and real link to destination
+    os.execute("ln -sf "..orig.." "..config.INSTALL.TOP.."/"..linkpath)
+    cache[package].links:write(linkpath.."\n")
+  end
 end
 
 ---
